backend/requirements.txt
# ================================
# AstroSource Pro - Backend Dependencies
# ================================

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic-settings==2.1.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS & Security
python-cors==1.7.0

# Astronomical Calculations
pyephem==4.1.5
swisseph==2.10.03.2
kerykeion==4.12.0

# Data Processing
numpy==1.25.2
pandas==2.1.4

# Image & SVG Generation
Pillow==10.1.0
cairosvg==2.7.1
svglib==1.5.1

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# WebSocket Support
websockets==12.0

# AI & OpenAI (optional)
openai==1.3.7

# Development Dependencies (comment out for production)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production Dependencies
gunicorn==21.2.0
